\docType{package}
\name{rtnc}
\alias{rtnc}
\alias{rtnc-package}
\title{rtnc: the Truncated Newton Constrained algorithm for R.}
\description{
  This package is a thin wrapper around Jean Sebastien
  Roy's C implementation of the Truncated Newton
  Constrained algorithm. It makes it possible to minimize a
  function under constraints, via the function
  \code{\link{tnc}}
}

